+++ cilium/tests/legacy/test_cilium.py
@@ -5,7 +5,7 @@
 from ..common import CILIUM_VERSION, requires_legacy_environment
 
 # Legacy common metrics
-from .common import ADDL_AGENT_METRICS, AGENT_DEFAULT_METRICS, OPERATOR_AWS_METRICS, OPERATOR_METRICS
+from .legacy_common import ADDL_AGENT_METRICS, AGENT_DEFAULT_METRICS, OPERATOR_AWS_METRICS, OPERATOR_METRICS
 
 pytestmark = [requires_legacy_environment]
