+++ cilium/tests/legacy/test_e2e.py
@@ -7,7 +7,7 @@ import pytest
 from ..common import requires_legacy_environment
 
 # Legacy common metrics
-from .common import AGENT_DEFAULT_METRICS, OPERATOR_METRICS
+from .legacy_common import AGENT_DEFAULT_METRICS, OPERATOR_METRICS
 
 pytestmark = [requires_legacy_environment]
